requirements.txt
pytz==2019.3
pandas_market_calendar==4.1.4
pandas==1.5.3
pytest==7.2.1